requirements.txt
streamlit
python-dotenv
langchain_google_genai
langchain_community
pypdf
PyPDF2
faiss-cpu
langchain-groq
groq
langchain